+++ server/src-lib/Hasura/GraphQL/Transport/HTTP.hs
@@ -17,6 +17,7 @@ import           Hasura.Server.Version                  (HasVersion)
 
 import qualified Data.Aeson                             as J
 import qualified Data.HashMap.Strict                    as Map
+import qualified Data.Sequence.NonEmpty                 as NESeq
 import qualified Database.PG.Query                      as Q
 import qualified Hasura.GraphQL.Execute                 as E
 import qualified Hasura.GraphQL.Execute.Query           as EQ
@@ -47,27 +48,30 @@ runGQ reqId userInfo reqHdrs req = do
         case queryPlan of
           E.ExecStepDB txGenSql -> do
             (telemTimeIO, telemQueryType, resp) <- runQueryDB reqId req userInfo txGenSql
-            return (telemCacheHit, Telem.Local, (telemTimeIO, telemQueryType, HttpResponse resp Nothing))
+            return (telemCacheHit, Telem.Local, (telemTimeIO, telemQueryType, HttpResponse resp []))
           E.ExecStepRemote (_name, (rsi, opDef)) ->
             runRemoteGQ telemCacheHit rsi opDef
           E.ExecStepRaw (name, json) -> do
             (telemTimeIO, obj) <- withElapsedTime $ do
               return $ encJFromJValue $ J.Object $ Map.singleton (G.unName name) json
-            return (telemCacheHit, Telem.Local, (telemTimeIO, Telem.Query, HttpResponse obj Nothing))
+            return (telemCacheHit, Telem.Local, (telemTimeIO, Telem.Query, HttpResponse obj []))
       E.MutationExecutionPlan mutationPlan -> do
         case mutationPlan of
           E.ExecStepDB tx -> do
             (telemTimeIO, telemQueryType, resp) <- runMutationDB reqId req userInfo tx
-            return (telemCacheHit, Telem.Local, (telemTimeIO, telemQueryType, HttpResponse resp Nothing))
+            return (telemCacheHit, Telem.Local, (telemTimeIO, telemQueryType, HttpResponse resp []))
           E.ExecStepRemote (_name, (rsi, opDef)) ->
             runRemoteGQ telemCacheHit rsi opDef
           E.ExecStepRaw (name, json) -> do
             (telemTimeIO, obj) <- withElapsedTime $ do
               return $ encJFromJValue $ J.Object $ Map.singleton (G.unName name) json
-            return (telemCacheHit, Telem.Local, (telemTimeIO, Telem.Query, HttpResponse obj Nothing))
+            return (telemCacheHit, Telem.Local, (telemTimeIO, Telem.Query, HttpResponse obj []))
       E.SubscriptionExecutionPlan _sub -> do
         throw400 UnexpectedPayload "subscriptions are not supported over HTTP, use websockets instead"
 {-
+      E.GExPHasura resolvedOp -> do
+        (telemTimeIO, telemQueryType, respHdrs, resp) <- runHasuraGQ reqId req userInfo resolvedOp
+        return (telemCacheHit, Telem.Local, (telemTimeIO, telemQueryType, HttpResponse resp respHdrs))
       E.GExPRemote rsi opDef  -> do
         let telemQueryType | G._todType opDef == G.OperationTypeMutation = Telem.Mutation
                             | otherwise = Telem.Query
@@ -106,7 +110,7 @@ runGQBatched reqId userInfo reqHdrs reqs =
       -- responses with distinct headers, so just do the simplest thing
       -- in this case, and don't forward any.
       let removeHeaders =
-            flip HttpResponse Nothing
+            flip HttpResponse []
             . encJFromList
             . map (either (encJFromJValue . encodeGQErr False) _hrBody)
           try = flip catchError (pure . Left) . fmap Right
@@ -169,7 +173,7 @@ runHasuraGQ
   -> GQLReqUnparsed
   -> UserInfo
   -> E.ExecOp
-  -> m (DiffTime, Telem.QueryType, EncJSON)
+  -> m (DiffTime, Telem.QueryType, HTTP.ResponseHeaders, EncJSON)
   -- ^ Also return 'Mutation' when the operation was a mutation, and the time
   -- spent in the PG query; for telemetry.
 runHasuraGQ reqId query userInfo resolvedOp = do
@@ -178,16 +182,16 @@ runHasuraGQ reqId query userInfo resolvedOp = do
     E.ExOpQuery tx genSql  -> do
       -- log the generated SQL and the graphql query
       L.unLogger logger $ QueryLog query genSql reqId
-      runLazyTx' pgExecCtx tx
-    E.ExOpMutation tx -> do
+      ([],) <$> runLazyTx' pgExecCtx tx
+    E.ExOpMutation respHeaders tx -> do
       -- log the graphql query
       L.unLogger logger $ QueryLog query Nothing reqId
-      runLazyTx pgExecCtx Q.ReadWrite $ withUserInfo userInfo tx
+      (respHeaders,) <$> runLazyTx pgExecCtx Q.ReadWrite (withUserInfo userInfo tx)
     E.ExOpSubs _ ->
       throw400 UnexpectedPayload
       "subscriptions are not supported over HTTP, use websockets instead"
-  resp <- liftEither respE
+  (respHdrs, resp) <- liftEither respE
   let !json = encodeGQResp $ GQSuccess $ encJToLBS resp
       telemQueryType = case resolvedOp of E.ExOpMutation{} -> Telem.Mutation ; _ -> Telem.Query
-  return (telemTimeIO, telemQueryType, json)
+  return (telemTimeIO, telemQueryType, respHdrs, json)
 -}