+++ server/src-lib/Hasura/Server/App.hs
@@ -20,6 +20,7 @@ import           Web.Spock.Core                         ((<//>))
 
 import qualified Control.Concurrent.Async.Lifted.Safe   as LA
 import qualified Data.ByteString.Lazy                   as BL
+import qualified Data.CaseInsensitive                   as CI
 import qualified Data.HashMap.Strict                    as M
 import qualified Data.HashSet                           as S
 import qualified Data.Text                              as T
@@ -197,6 +198,10 @@ buildQCtx = do
   sqlGenCtx <- scSQLGenCtx . hcServerCtx <$> ask
   return $ QCtx userInfo cache sqlGenCtx
 
+setHeader :: MonadIO m => HTTP.Header -> Spock.ActionT m ()
+setHeader (headerName, headerValue) =
+  Spock.setHeader (bsToTxt $ CI.original headerName) (bsToTxt headerValue)
+
 -- | Typeclass representing the metadata API authorization effect
 class MetadataApiAuthorization m where
   authorizeMetadataApi :: RQLQuery -> UserInfo -> Handler m ()
@@ -269,21 +274,19 @@ mkSpockAction serverCtx qErrEncoder qErrModifier apiHandler = do
         case result of
           JSONResp (HttpResponse encJson h) ->
             possiblyCompressedLazyBytes userInfo reqId req reqBody qTime (encJToLBS encJson)
-              (pure jsonHeader <> mkHeaders h) reqHeaders
+              (pure jsonHeader <> h) reqHeaders
           RawResp (HttpResponse rawBytes h) ->
-            possiblyCompressedLazyBytes userInfo reqId req reqBody qTime rawBytes (mkHeaders h) reqHeaders
+            possiblyCompressedLazyBytes userInfo reqId req reqBody qTime rawBytes h reqHeaders
 
       possiblyCompressedLazyBytes userInfo reqId req reqBody qTime respBytes respHeaders reqHeaders = do
         let (compressedResp, mEncodingHeader, mCompressionType) =
               compressResponse (Wai.requestHeaders req) respBytes
             encodingHeader = maybe [] pure mEncodingHeader
-            reqIdHeader = (requestIdHeader, unRequestId reqId)
+            reqIdHeader = (requestIdHeader, txtToBs $ unRequestId reqId)
             allRespHeaders = pure reqIdHeader <> encodingHeader <> respHeaders
         lift $ logHttpSuccess logger userInfo reqId req reqBody respBytes compressedResp qTime mCompressionType reqHeaders
-        mapM_ (uncurry Spock.setHeader) allRespHeaders
+        mapM_ setHeader allRespHeaders
         Spock.lazyBytes compressedResp
-
-      mkHeaders = maybe [] (map unHeader)
 
 v1QueryHandler
   :: (HasVersion, MonadIO m, MonadBaseControl IO m, MetadataApiAuthorization m)
@@ -295,7 +298,7 @@ v1QueryHandler query = do
   logger <- scLogger . hcServerCtx <$> ask
   res <- bool (fst <$> dbAction) (withSCUpdate scRef logger dbAction) $
          queryModifiesSchemaCache query
-  return $ HttpResponse res Nothing
+  return $ HttpResponse res []
   where
     -- Hit postgres
     dbAction = do
@@ -340,14 +343,14 @@ v1GQHandler = v1Alpha1GQHandler
 --   sqlGenCtx <- scSQLGenCtx . hcServerCtx <$> ask
 --   enableAL <- scEnableAllowlist . hcServerCtx <$> ask
 --   res <- GE.explainGQLQuery pgExecCtx sc sqlGenCtx enableAL query
---   return $ HttpResponse res Nothing
+--   return $ HttpResponse res []
 
 v1Alpha1PGDumpHandler :: (MonadIO m) => PGD.PGDumpReqBody -> Handler m APIResp
 v1Alpha1PGDumpHandler b = do
   onlyAdmin
   ci <- scConnInfo . hcServerCtx <$> ask
   output <- PGD.execPGDump b ci
-  return $ RawResp $ HttpResponse output (Just [Header sqlHeader])
+  return $ RawResp $ HttpResponse output [sqlHeader]
 
 consoleAssetsHandler
   :: (MonadIO m, HttpLog m)
@@ -365,7 +368,7 @@ consoleAssetsHandler logger dir path = do
   either (onError reqHeaders) onSuccess eFileContents
   where
     onSuccess c = do
-      mapM_ (uncurry Spock.setHeader) headers
+      mapM_ setHeader headers
       Spock.lazyBytes c
     onError :: (MonadIO m, HttpLog m) => [HTTP.Header] -> IOException -> Spock.ActionT m ()
     onError hdrs = raiseGenericApiError logger hdrs . err404 NotFound . T.pack . show
@@ -374,7 +377,7 @@ consoleAssetsHandler logger dir path = do
     (fileName, encHeader) = case T.stripSuffix ".gz" fn of
       Just v  -> (v, [gzipHeader])
       Nothing -> (fn, [])
-    mimeType = bsToTxt $ defaultMimeLookup fileName
+    mimeType = defaultMimeLookup fileName
     headers = ("Content-Type", mimeType) : encHeader
 
 class (Monad m) => ConsoleRenderer m where
@@ -551,7 +554,7 @@ httpApp corsCfg serverCtx enableConsole consoleAssetsDir enableTelemetry = do
         else Spock.setStatus HTTP.status500 >> Spock.text "ERROR"
 
     Spock.get "v1/version" $ do
-      uncurry Spock.setHeader jsonHeader
+      setHeader jsonHeader
       Spock.lazyBytes $ encode $ object [ "version" .= currentVersion ]
 
     when enableMetadata $ do
@@ -577,7 +580,7 @@ httpApp corsCfg serverCtx enableConsole consoleAssetsDir enableTelemetry = do
         mkGetHandler $ do
           onlyAdmin
           let res = encJFromJValue $ runGetConfig (scAuthMode serverCtx)
-          return $ JSONResp $ HttpResponse res Nothing
+          return $ JSONResp $ HttpResponse res []
 
     when enableGraphQL $ do
       Spock.post "v1alpha1/graphql" $ spockAction GH.encodeGQErr id $
@@ -591,22 +594,22 @@ httpApp corsCfg serverCtx enableConsole consoleAssetsDir enableTelemetry = do
         mkGetHandler $ do
           onlyAdmin
           respJ <- liftIO $ EKG.sampleAll $ scEkgStore serverCtx
-          return $ JSONResp $ HttpResponse (encJFromJValue $ EKG.sampleToJson respJ) Nothing
+          return $ JSONResp $ HttpResponse (encJFromJValue $ EKG.sampleToJson respJ) []
       Spock.get "dev/plan_cache" $ spockAction encodeQErr id $
         mkGetHandler $ do
           onlyAdmin
           respJ <- liftIO $ E.dumpPlanCache $ scPlanCache serverCtx
-          return $ JSONResp $ HttpResponse (encJFromJValue respJ) Nothing
+          return $ JSONResp $ HttpResponse (encJFromJValue respJ) []
       Spock.get "dev/subscriptions" $ spockAction encodeQErr id $
         mkGetHandler $ do
           onlyAdmin
           respJ <- liftIO $ EL.dumpLiveQueriesState False $ scLQState serverCtx
-          return $ JSONResp $ HttpResponse (encJFromJValue respJ) Nothing
+          return $ JSONResp $ HttpResponse (encJFromJValue respJ) []
       Spock.get "dev/subscriptions/extended" $ spockAction encodeQErr id $
         mkGetHandler $ do
           onlyAdmin
           respJ <- liftIO $ EL.dumpLiveQueriesState True $ scLQState serverCtx
-          return $ JSONResp $ HttpResponse (encJFromJValue respJ) Nothing
+          return $ JSONResp $ HttpResponse (encJFromJValue respJ) []
 
     forM_ [Spock.GET, Spock.POST] $ \m -> Spock.hookAny m $ \_ -> do
       req <- Spock.request
@@ -671,6 +674,6 @@ raiseGenericApiError logger headers qErr = do
   reqBody <- liftIO $ Wai.strictRequestBody req
   reqId <- getRequestId $ Wai.requestHeaders req
   lift $ logHttpError logger Nothing reqId req (Left reqBody) qErr headers
-  uncurry Spock.setHeader jsonHeader
+  setHeader jsonHeader
   Spock.setStatus $ qeStatus qErr
   Spock.lazyBytes $ encode qErr